+++ src/com/garbagemule/MobArena/health/HealthStrategyStandard.java
@@ -6,6 +6,7 @@ public class HealthStrategyStandard implements HealthStrategy
 {
     @Override
     public void setHealth(Player p, double health) {
-        p.setHealth(health);
+        // TODO: Remove cast for 1.6
+        p.setHealth((int) health);
     }
 }
